[bees] Refactor TaskStore Persistence and Decouple from Ticket Model
## What
Decoupled the `TaskStore` from the `Ticket` domain model in the `bees` package to enforce a clean repository pattern. Moved `TaskStore` to its own file and removed legacy `save` methods from `Ticket`.

## Why
To improve separation of concerns and make the codebase more maintainable by centralizing persistence logic in the repository (`TaskStore`) rather than the domain model (`Ticket`).

## Changes
- **Core**:
  - Extracted `TaskStore` and helpers from `ticket.py` to a new `task_store.py`.
  - Removed `TaskStore` and `save`/`save_metadata` methods from `Ticket` in `ticket.py`.
  - Updated `bees/__init__.py` to export `TaskStore` from the new location.
- **Callers**:
  - Updated `playbook.py`, `scheduler.py`, `chat.py`, `server.py`, and `edit_tags.py` to use `TaskStore` or `task_store` instance for saving metadata instead of `ticket.save_metadata()`.
  - Updated `run_event_hooks` in `playbook.py` to accept a `store` argument to allow hooks to persist changes.
- **Tests**:
  - Updated `test_chat.py`, `test_tasks.py`, `test_scheduler.py`, and `test_playbook.py` to use `TaskStore` and pass it to hooks where needed.
- **Configuration**:
  - Added a dummy `test:python` script to `packages/bees/hivetool/package.json` to avoid errors when running workspace tests.

## Testing
Ran python tests in `packages/bees` using `npm run test:python -w packages/bees`. All 161 tests passed.
