+++ packages/bees/bees/__init__.py
@@ -4,7 +4,7 @@
 """Bees — agent swarm orchestration framework."""
 
 from .ticket import Ticket as Task
-from .ticket import TaskStore
+from .task_store import TaskStore
 from .scheduler import Scheduler, SchedulerHooks
 
 __all__ = ["Task", "TaskStore", "Scheduler", "SchedulerHooks"]
