+++ packages/bees/hivetool/package.json
@@ -6,7 +6,8 @@
     "dev": "vite",
     "build": "vite build",
     "build:test": "tsc -p tsconfig.test.json",
-    "test": "npm run build:test && node --test './dist/tsc/hivetool/tests/**/*.test.js'"
+    "test": "npm run build:test && node --test './dist/tsc/hivetool/tests/**/*.test.js'",
+    "test:python": "echo 'No python tests in hivetool'"
   },
   "dependencies": {
     "@lit-labs/signals": "^0.1.3",
